pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gclass-cli"
version = "0.3.0"
description = "Google Classroom CLI"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "mrgolddev", email = "mrgolddev@gmail.com"}]
requires-python = ">=3.10"
dependencies = [
  "click>=8.1",
  "google-api-python-client>=2.131",
  "google-auth-oauthlib>=1.2",
  "google-auth-httplib2>=0.2.0",
  "platformdirs>=4.2",
  "rich>=13.7"
]

[project.scripts]
gclass = "gclass.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
